+++ app/src/main/kotlin/com/numero/sojodia/repository/BusDataRepository.kt
@@ -1,6 +1,5 @@
 package com.numero.sojodia.repository
 
-import android.util.Log
 import com.numero.sojodia.resource.IBusDataSource
 import com.numero.sojodia.resource.datasource.BusTime
 import com.numero.sojodia.resource.model.BusDataResponse
@@ -33,13 +32,10 @@ class BusDataRepository(
     override fun loadAndSaveBusData(): Observable<BusDataResponse> = busDataSource.getAndSaveBusData()
 
     override fun reloadBusData() {
-        val start = System.currentTimeMillis()
         val list = busDataSource.loadAllBusTimeObservable().blockingGet()
         tkBusTimeListGoing = list.asSequence().filter { it.routeId == Route.TkToKutc.id }.toList()
         tkBusTimeListReturn = list.asSequence().filter { it.routeId == Route.KutcToTk.id }.toList()
         tndBusTimeListGoing = list.asSequence().filter { it.routeId == Route.TndToKutc.id }.toList()
         tndBusTimeListReturn = list.asSequence().filter { it.routeId == Route.KutcToTnd.id }.toList()
-
-        Log.d("Log", "time : ${System.currentTimeMillis() - start}")
     }
 }
